+++ networking/eth2/src/main/java/tech/pegasys/teku/networking/eth2/rpc/beaconchain/methods/BeaconBlocksByRootMessageHandler.java
@@ -124,7 +124,7 @@ public class BeaconBlocksByRootMessageHandler
     final UInt64 currentEpoch = recentChainData.getCurrentEpoch().orElse(UInt64.ZERO);
     final SpecMilestone milestone = spec.getForkSchedule().getSpecMilestoneAtEpoch(currentEpoch);
     return milestone.isGreaterThanOrEqualTo(SpecMilestone.DENEB)
-        ? spec.getNetworkingConfigDeneb().orElseThrow().getMaxRequestBlocksDeneb()
+        ? UInt64.valueOf(spec.getNetworkingConfigDeneb().orElseThrow().getMaxRequestBlocksDeneb())
         : UInt64.valueOf(spec.getNetworkingConfig().getMaxRequestBlocks());
   }
