+++ networking/eth2/src/main/java/tech/pegasys/teku/networking/eth2/rpc/beaconchain/methods/BlobSidecarsByRangeMessageHandler.java
@@ -102,8 +102,7 @@ public class BlobSidecarsByRangeMessageHandler
         SpecConfigDeneb.required(spec.atSlot(endSlot).getConfig());
     final UInt64 maxBlobsPerBlock = UInt64.valueOf(specConfigDeneb.getMaxBlobsPerBlock());
     final UInt64 maxRequestBlobSidecars =
-        specConfigDeneb.getNetworkingConfig().getMaxRequestBlobSidecars();
-
+        UInt64.valueOf(specConfigDeneb.getNetworkingConfig().getMaxRequestBlobSidecars());
     final UInt64 requestedCount = message.getCount().times(maxBlobsPerBlock);
 
     if (requestedCount.isGreaterThan(maxRequestBlobSidecars)) {
